Changed board rendering behaviour
The previous rendering method, which created a sprite for each tile and
updated their properties (such as position) on every move, as well as
drawing all of them all of the time, proved to be too much for the Wii
to handle at large board sizes.

Instead, the Game class itself keeps all of the attributes that can
change and draws only the tiles that are presented on screen. This
reduces complexity and object count, and significantly speeds drawing.